move statsArea to own file
