+++ src/app/(post-login)/submission/program/[shortName]/clinical-submission/components/FilesNavigator/FileRecordTable.tsx
@@ -19,86 +19,17 @@
 
 import {
   CellContentCenter,
-  DataTableStarIcon,
-  StatArea as StatAreaDisplay,
   SubmissionInfoArea,
   TableInfoHeaderContainer,
 } from "@/app/components/Table/common";
+import { toDisplayRowIndex } from "@/global/utils";
 import { css } from "@/lib/emotion";
-import { Icon, Table, useTheme } from "@icgc-argo/uikit";
+import { Table, useTheme } from "@icgc-argo/uikit";
 import get from "lodash/get";
-import pluralize from "pluralize";
 import { CSSProperties, ComponentProps, createRef } from "react";
 import { ClinicalSubmissionEntity } from "../../types";
-
-const StarIcon = DataTableStarIcon;
-
-type RecordState = "NEW" | "NONE" | "UPDATED" | "ERROR" | "WARNING";
-
-type FileStat = {
-  newCount: number;
-  noUpdateCount: number;
-  updateCount: number;
-  errorCount: number;
-};
-
-export const FILE_STATE_COLORS: {
-  [k in RecordState]: ComponentProps<typeof StarIcon>["fill"];
-} = {
-  ERROR: "error",
-  WARNING: "warning",
-  NEW: "accent2",
-  NONE: "grey_1",
-  UPDATED: "accent3_dark",
-};
-
-const StatsArea = ({
-  fileStat,
-  total,
-  isSubmissionValidated,
-}: {
-  fileStat: FileStat;
-  total: number;
-  isSubmissionValidated: boolean;
-}) => {
-  return (
-    <StatAreaDisplay.Container>
-      <StatAreaDisplay.Section>
-        {total.toLocaleString()} Total
-      </StatAreaDisplay.Section>
-      <StatAreaDisplay.Section faded={!isSubmissionValidated}>
-        <Icon name="chevron_right" fill="grey_1" width="8px" />
-      </StatAreaDisplay.Section>
-      <StatAreaDisplay.Section faded={!isSubmissionValidated}>
-        <StatAreaDisplay.StatEntryContainer>
-          <StatAreaDisplay.StarIcon fill={FILE_STATE_COLORS.ERROR} />
-          {isSubmissionValidated && fileStat.errorCount.toLocaleString()}{" "}
-          {pluralize("Error", fileStat.errorCount)}
-        </StatAreaDisplay.StatEntryContainer>
-      </StatAreaDisplay.Section>
-      <StatAreaDisplay.Section faded={!isSubmissionValidated}>
-        <StatAreaDisplay.StatEntryContainer>
-          <StatAreaDisplay.StarIcon fill={FILE_STATE_COLORS.UPDATED} />
-          {isSubmissionValidated && fileStat.updateCount.toLocaleString()}{" "}
-          Updated
-        </StatAreaDisplay.StatEntryContainer>
-      </StatAreaDisplay.Section>
-      <StatAreaDisplay.Section faded={!isSubmissionValidated}>
-        <StatAreaDisplay.StatEntryContainer>
-          <StatAreaDisplay.StarIcon fill={FILE_STATE_COLORS.NEW} />
-          {isSubmissionValidated && fileStat.newCount.toLocaleString()} New
-        </StatAreaDisplay.StatEntryContainer>
-      </StatAreaDisplay.Section>
-      <StatAreaDisplay.Section faded={!isSubmissionValidated}>
-        <StatAreaDisplay.StatEntryContainer>
-          <StatAreaDisplay.StarIcon fill={FILE_STATE_COLORS.NONE} />
-          {isSubmissionValidated && fileStat.noUpdateCount.toLocaleString()} No
-          Update
-        </StatAreaDisplay.StatEntryContainer>
-      </StatAreaDisplay.Section>
-    </StatAreaDisplay.Container>
-  );
-};
+import { FILE_STATE_COLORS } from "../SummaryTable";
+import StatsArea from "./StatsArea";
 
 const REQUIRED_FILE_ENTRY_FIELDS = {
   ROW: "row",
